+++ RFRuntime.h
@@ -89,3 +89,22 @@ NSString *footerViewFramePath = @keypathClassInstance(UITableView, tableFooterVi
 #define keypathClassInstance2(CLASS, PROPERTY, PATH)\
     (({CLASS *_proxy_; ((void)(NO && ((void)_proxy_.PROPERTY.PATH, NO)), # PATH);}))
 
+#pragma mark - Language Addition
+
+/** Define a const NSString
+ 
+ @code
+ 
+ RFDefineConstString(aaa);
+ RFDefineConstString(bbb, "something");
+ 
+ // Will be
+ NSString *const aaa = @"aaa";
+ NSString *const bbb = @"something";
+ 
+ @endcode
+ */
+#define RFDefineConstString(...) metamacro_if_eq(1, metamacro_argcount(__VA_ARGS__))(_RFDefineConstString1(metamacro_at0(__VA_ARGS__)))(_RFDefineConstString2(metamacro_at0(__VA_ARGS__), metamacro_at1(__VA_ARGS__)))
+
+#define _RFDefineConstString1(name) NSString *const name = @metamacro_stringify_(name)
+#define _RFDefineConstString2(name, value) NSString *const name = @value